requirements.txt
streamlit
streamlit-webrtc
torch
torchvision
pillow
av
opencv-python-headless
numpy
matplotlib